Some Code Cleanup and minor improvements
