+++ MaterialSkin/ColorHelper.cs
@@ -10,6 +10,52 @@ namespace MaterialSkin
     public static class ColorHelper
     {
         /// <summary>
+        /// Tints the color by the given percent.
+        /// </summary>
+        /// <param name="color">The color being tinted.</param>
+        /// <param name="percent">The percent to tint. Ex: 0.1 will make the color 10% lighter.</param>
+        /// <returns>The new tinted color.</returns>
+        public static Color Lighten(this Color color, float percent)
+        {
+            var lighting = color.GetBrightness();
+            lighting = lighting + lighting * percent;
+            if (lighting > 1.0)
+            {
+                lighting = 1;
+            }
+            else if (lighting <= 0)
+            {
+                lighting = 0.1f;
+            }
+            var tintedColor = ColorHelper.FromHsl(color.A, color.GetHue(), color.GetSaturation(), lighting);
+
+            return tintedColor;
+        }
+
+        /// <summary>
+        /// Tints the color by the given percent.
+        /// </summary>
+        /// <param name="color">The color being tinted.</param>
+        /// <param name="percent">The percent to tint. Ex: 0.1 will make the color 10% darker.</param>
+        /// <returns>The new tinted color.</returns>
+        public static Color Darken(this Color color, float percent)
+        {
+            var lighting = color.GetBrightness();
+            lighting = lighting - lighting * percent;
+            if (lighting > 1.0)
+            {
+                lighting = 1;
+            }
+            else if (lighting <= 0)
+            {
+                lighting = 0;
+            }
+            var tintedColor = ColorHelper.FromHsl(color.A, color.GetHue(), color.GetSaturation(), lighting);
+
+            return tintedColor;
+        }
+
+        /// <summary>
         /// Converts the HSL values to a Color.
         /// </summary>
         /// <param name="alpha">The alpha.</param>
