+++ MaterialSkin/Controls/MaterialCheckedListBox.cs
@@ -25,13 +25,11 @@
 
         public MaterialCheckedListBox() : base()
         {
-            InitializeComponent();
             this.DoubleBuffered = true;
             this.Items = new ItemsList(this);
             this.AutoScroll = true;
             BackColorChanged += (sender, args) => BackColor = SkinManager.GetControlBackgroundColor();
             ForeColorChanged += (sender, args) => ForeColor = SkinManager.GetPrimaryTextColor();
-            
         }
